+++ src/main/java/uk/ashleybye/tictactoe/game/Board.java
@@ -1,4 +1,4 @@
-package uk.ashleybye.tictactoe;
+package uk.ashleybye.tictactoe.game;
 
 import java.util.ArrayList;
 import java.util.Arrays;
