+++ src/main/java/uk/ashleybye/tictactoe/game/GameState.java
@@ -1,4 +1,4 @@
-package uk.ashleybye.tictactoe;
+package uk.ashleybye.tictactoe.game;
 
 public enum GameState {
   PLAYING, GAME_OVER, READY
