+++ src/main/java/uk/ashleybye/tictactoe/game/HumanMovePublisher.java
@@ -1,4 +1,4 @@
-package uk.ashleybye.tictactoe;
+package uk.ashleybye.tictactoe.game;
 
 public interface HumanMovePublisher {
