+++ src/main/java/uk/ashleybye/tictactoe/game/HumanPlayer.java
@@ -1,4 +1,4 @@
-package uk.ashleybye.tictactoe;
+package uk.ashleybye.tictactoe.game;
 
 public class HumanPlayer implements Player, HumanMoveSubscriber {
