+++ src/main/java/uk/ashleybye/tictactoe/game/Mark.java
@@ -1,4 +1,4 @@
-package uk.ashleybye.tictactoe;
+package uk.ashleybye.tictactoe.game;
 
 interface Mark {
