+++ src/main/java/uk/ashleybye/tictactoe/game/Player.java
@@ -1,4 +1,4 @@
-package uk.ashleybye.tictactoe;
+package uk.ashleybye.tictactoe.game;
 
 public interface Player {
