+++ src/test/java/uk/ashleybye/tictactoe/game/HumanPlayerTest.java
@@ -1,4 +1,4 @@
-package uk.ashleybye.tictactoe;
+package uk.ashleybye.tictactoe.game;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
