+++ src/test/java/uk/ashleybye/tictactoe/game/MockGame.java
@@ -1,4 +1,4 @@
-package uk.ashleybye.tictactoe;
+package uk.ashleybye.tictactoe.game;
 
 import java.util.Arrays;
 import java.util.List;
