+++ src/test/java/uk/ashleybye/tictactoe/game/TestHelpers.java
@@ -1,4 +1,4 @@
-package uk.ashleybye.tictactoe;
+package uk.ashleybye.tictactoe.game;
 
 public class TestHelpers {
